Moving functions from server to api to dodge circular import issues
